renamed classnames and cleaned css code
