+++ homestead/src/family_pages/FamilyPages.css
@@ -1,0 +1,255 @@
+.familypages_container {
+    margin-top: 150px;
+    height: 100%;
+    width: 100%;
+    display: flex;
+    background-image: url("https://htmlcolorcodes.com/assets/images/colors/pastel-blue-color-solid-background-1920x1080.png");
+}
+
+.sidebars {
+    flex: 0.25;
+    margin-left: 30px;
+    margin-top: 30px;
+}
+
+/* FIRST SIDEBAR */
+
+.sidebars_first>img {
+    height: 290px;
+    width: 290px;
+    border-radius: 50%;
+    /*rounds radius*/
+    margin: 30px auto;
+    display: flex;
+    /*expands items to fill available free space or shrinks them to prevent overflow*/
+}
+
+.sidebars_first>h1 {
+    font-family: var(--font-form);
+    font-size: 22px;
+    text-align: center;
+    color: var(--background-color);
+}
+
+.sidebars_first {
+    background-color: rgba(25, 71, 97, 0.5);
+    border-radius: 30px;
+    padding: 30px;
+}
+
+.score {
+    display: flex;
+    justify-content: space-around;
+    text-align: center;
+    background-color: var(--background-color);
+    padding: 10px;
+    border-radius: 30px;
+    margin-top: 30px;
+    margin-bottom: 30px;
+    margin-left: 30px;
+    margin-right: 30px;
+}
+
+.service {
+    font-family: var(--oswaldo);
+    font-weight: bold;
+}
+
+.fellowship {
+    font-family: var(--oswaldo);
+    font-weight: bold;
+}
+
+.leadership {
+    font-family: var(--oswaldo);
+}
+
+.service>h2 {
+    color: var(--service-color);
+    font-weight: 900;
+}
+
+.fellowship>h2 {
+    color: var(--fellowship-color);
+}
+
+.leadership>h2 {
+    color: var(--leadership-color);
+}
+
+.bigsLittles_container {
+    display: flex;
+}
+
+.bigsLittles_container>button {
+    font-family: Oswald;
+    font-style: normal;
+    font-weight: bold;
+    font-size: 24px;
+    line-height: 36px;
+    flex: 0.5;
+    text-align: center;
+    color: #FFFFFF;
+    background-color: Transparent;
+    background-repeat: no-repeat;
+    border: none;
+    cursor: pointer;
+    overflow: hidden;
+    outline: none;
+    padding: none;
+    transition: all .1s ease-in-out;
+}
+
+.bigLittleButton:hover {
+    text-shadow: 2px 2px 5px black;
+}
+
+.bigsLittles_container>h5 {
+    flex: 0.5;
+    font-family: Oswald;
+    font-weight: bolder;
+    font-size: 24px;
+    line-height: 36x;
+    text-align: center;
+    color: #FFFFFF;
+}
+
+.bigLittle_card {
+    display: flex;
+    justify-content: space-around;
+    background-color: var(--background-color);
+    padding: 10px;
+    border-radius: 30px;
+    margin-top: 30px;
+    margin-bottom: 30px;
+    margin-left: 30px;
+    margin-right: 30px;
+}
+
+.bigLittle_card>img {
+    border-radius: 50%;
+    width: 122px;
+    height: 122px;
+}
+
+.nameEmail>h1 {
+    margin-top: 20px;
+    font-family: var(--font-form);
+    font-weight: bold;
+    font-size: 24px;
+    line-height: 30px;
+    text-align: center;
+    color: #000000;
+}
+
+.nameEmail>h2 {
+    font-family: var(--font-form);
+    font-weight: normal;
+    font-size: 20px;
+    line-height: 30px;
+    text-align: center;
+    color: #000000;
+}
+
+.spotify {
+    background-color: rgba(25, 71, 97, 0.5);
+    border-radius: 30px;
+    padding: 30px;
+    margin-top: 30px;
+}
+
+.spotify>img {
+    height: 641px;
+    width: 472px;
+    margin: 30px auto;
+    display: flex;
+}
+
+.spotify>h1 {
+    font-family: var(--font-form);
+    font-style: normal;
+    font-weight: bold;
+    font-size: 22px;
+    text-align: center;
+    color: #FFFFFF;
+}
+
+/* SECOND SIDEBAR */
+
+.sidebars_second {
+    flex: 0.9;
+    margin-left: 100px;
+    margin-top: 30px;
+}
+
+.family_picture {
+    border-radius: 50px;
+    width: 100%;
+    height: 371px;
+    object-fit: cover;
+}
+
+.welcome>h1 {
+    margin-top: 30px;
+    margin-bottom: 30px;
+    font-weight: 500px;
+    font-size: 36px;
+    line-height: 53px;
+    letter-spacing: 0.02em;
+    font-family: var(--oswaldo);
+    color: #194761;
+}
+
+.bio>h1 {
+    font-family: var(--font-form);
+    font-style: normal;
+    font-weight: normal;
+    font-size: 22px;
+    line-height: 30px;
+    color: #000000;
+    margin-bottom: 30px;
+}
+
+.memberBio {
+    margin-top: 30px;
+    margin-bottom: 30px;
+    font-family: var(--oswaldo);
+    font-weight: 500px;
+    font-size: 18px;
+    line-height: 53px;
+    letter-spacing: 0.02em;
+    color: #194761;
+}
+
+.member_card {
+    display: flex;
+    background: var(--background-color);
+    padding: 10px;
+    border-radius: 30px;
+    margin-bottom: 30px;
+}
+
+.member_nameImage>img {
+    border-radius: 50%;
+    width: 205px;
+    height: 205px;
+    padding: 30px;
+}
+
+.member_nameImage>h1 {
+    font-family: var(--font-form);
+    font-weight: bold;
+    font-size: 24px;
+    color: #000000;
+    margin-left: 30px;
+    margin-bottom: 30px;
+}
+
+.member_description>h2 {
+    padding: 30px;
+    background: var(--background-color);
+    border-radius: 10px;
+    font-weight: normal;
+    font-size: 20px;
+    line-height: 25px;
+}
